Made exampletest work and did some small refactoring
