+++ src/test/java/no/uib/inf112/desktop/ExampleTest.java
@@ -1,9 +1,14 @@
 package no.uib.inf112.desktop;
 
 import no.uib.inf112.core.RoboRally;
+import no.uib.inf112.core.player.Movement;
+import no.uib.inf112.core.player.Player;
+import no.uib.inf112.core.player.Robot;
+import no.uib.inf112.core.util.Vector2Int;
 import org.junit.Before;
 import org.junit.Test;
 
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 
 public class ExampleTest extends TestGraphics{
@@ -13,10 +18,15 @@ public class ExampleTest extends TestGraphics{
     public void setUp() {
         Main.HEADLESS = true;
         roboRally = new RoboRally();
+        roboRally.getPlayerHandler().generatePlayers(Main.HEADLESS);
     }
 
     @Test
-    public void exampleTest() {
-        assertNotNull(roboRally.getCurrentMap());
+    public void MoveShouldMovePlayer() {
+        Robot robot = roboRally.getPlayerHandler().mainPlayer().getRobot();
+        Vector2Int pos = new Vector2Int(((Robot) robot).getX(), robot.getY());
+
+        robot.move(Movement.MOVE_1);
+        assertNotEquals(pos, new Vector2Int(robot.getX(), robot.getY()));
     }
 }
